+++ src/engine.h
@@ -186,12 +186,12 @@ struct engine {
   integertime_t ti_beg_max;
 
   /* Number of particles updated in the previous step */
-  size_t updates, g_updates, s_updates;
+  long long updates, g_updates, s_updates;
 
   /* Number of updates since the last rebuild */
-  size_t updates_since_rebuild;
-  size_t g_updates_since_rebuild;
-  size_t s_updates_since_rebuild;
+  long long updates_since_rebuild;
+  long long g_updates_since_rebuild;
+  long long s_updates_since_rebuild;
 
   /* Properties of the previous step */
   int step_props;
